secretmanager/snippets-adk/agent_global/requirements.txt
google-adk>=1.29
